+++ tests/test_core.py
@@ -11,11 +11,10 @@ import platform
 import warnings
 
 # 3rd party
-import six
 import pytest
-from six.moves import reload_module
 
 # local
+from blessed._compat import StringIO
 from .accessories import TestTerminal, unicode_cap, as_subprocess
 from .conftest import IS_WINDOWS
 
@@ -23,6 +22,15 @@ try:
     from unittest import mock
 except ImportError:
     import mock
+
+
+# reload was a built-in in PY2, then moved to imp in PY3, then moved to importlib in PY34
+if sys.version_info[:2] >= (3, 4):
+    from importlib import reload as reload_module
+elif sys.version_info[0] >= 3:
+    from imp import reload as reload_module  # pylint: disable=deprecated-module
+else:
+    reload_module = reload  # pylint: disable=undefined-variable  # noqa: F821
 
 
 def test_export_only_Terminal():
@@ -35,7 +43,7 @@ def test_null_location(all_terms):
     """Make sure ``location()`` with no args just does position restoration."""
     @as_subprocess
     def child(kind):
-        t = TestTerminal(stream=six.StringIO(), force_styling=True)
+        t = TestTerminal(stream=StringIO(), force_styling=True)
         with t.location():
             pass
         expected_output = u''.join(
@@ -49,7 +57,7 @@ def test_location_to_move_xy(all_terms):
     """``location()`` and ``move_xy()`` receive complimentary arguments."""
     @as_subprocess
     def child(kind):
-        buf = six.StringIO()
+        buf = StringIO()
         t = TestTerminal(stream=buf, force_styling=True)
         x, y = 12, 34
         with t.location(y, x):
@@ -65,7 +73,7 @@ def test_yield_keypad():
     @as_subprocess
     def child(kind):
         # given,
-        t = TestTerminal(stream=six.StringIO(), force_styling=True)
+        t = TestTerminal(stream=StringIO(), force_styling=True)
         expected_output = u''.join((t.smkx, t.rmkx))
 
         # exercise,
@@ -83,7 +91,7 @@ def test_null_fileno():
     @as_subprocess
     def child():
         # This simulates piping output to another program.
-        out = six.StringIO()
+        out = StringIO()
         out.fileno = None
         t = TestTerminal(stream=out)
         assert (t.save == u'')
@@ -99,32 +107,32 @@ def test_number_of_colors_without_tty():
 
     @as_subprocess
     def child_256_nostyle():
-        t = TestTerminal(stream=six.StringIO())
+        t = TestTerminal(stream=StringIO())
         assert (t.number_of_colors == 0)
 
     @as_subprocess
     def child_256_forcestyle():
-        t = TestTerminal(stream=six.StringIO(), force_styling=True)
+        t = TestTerminal(stream=StringIO(), force_styling=True)
         assert (t.number_of_colors == 256)
 
     @as_subprocess
     def child_8_forcestyle():
         # 'ansi' on freebsd returns 0 colors. We use 'cons25', compatible with its kernel tty.c
         kind = 'cons25' if platform.system().lower() == 'freebsd' else 'ansi'
-        t = TestTerminal(kind=kind, stream=six.StringIO(),
+        t = TestTerminal(kind=kind, stream=StringIO(),
                          force_styling=True)
         assert (t.number_of_colors == 8)
 
     @as_subprocess
     def child_0_forcestyle():
-        t = TestTerminal(kind='vt220', stream=six.StringIO(),
+        t = TestTerminal(kind='vt220', stream=StringIO(),
                          force_styling=True)
         assert (t.number_of_colors == 0)
 
     @as_subprocess
     def child_24bit_forcestyle_with_colorterm():
         os.environ['COLORTERM'] = 'truecolor'
-        t = TestTerminal(kind='vt220', stream=six.StringIO(),
+        t = TestTerminal(kind='vt220', stream=StringIO(),
                          force_styling=True)
         assert (t.number_of_colors == 1 << 24)
 
@@ -163,7 +171,7 @@ def test_init_descriptor_always_initted(all_terms):
     """Test height and width with non-tty Terminals."""
     @as_subprocess
     def child(kind):
-        t = TestTerminal(kind=kind, stream=six.StringIO())
+        t = TestTerminal(kind=kind, stream=StringIO())
         assert t._init_descriptor == sys.__stdout__.fileno()
         assert (isinstance(t.height, int))
         assert (isinstance(t.width, int))
@@ -326,7 +334,7 @@ def test_yield_fullscreen(all_terms):
     """Ensure ``fullscreen()`` writes enter_fullscreen and exit_fullscreen."""
     @as_subprocess
     def child(kind):
-        t = TestTerminal(stream=six.StringIO(), force_styling=True)
+        t = TestTerminal(stream=StringIO(), force_styling=True)
         t.enter_fullscreen = u'BEGIN'
         t.exit_fullscreen = u'END'
         with t.fullscreen():
@@ -341,7 +349,7 @@ def test_yield_hidden_cursor(all_terms):
     """Ensure ``hidden_cursor()`` writes hide_cursor and normal_cursor."""
     @as_subprocess
     def child(kind):
-        t = TestTerminal(stream=six.StringIO(), force_styling=True)
+        t = TestTerminal(stream=StringIO(), force_styling=True)
         t.hide_cursor = u'BEGIN'
         t.normal_cursor = u'END'
         with t.hidden_cursor():